+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Other guiding principles:
 - **amaru-ledger**: reject treasury withdrawal proposals that reference unregistered reward accounts.  ([#1032][], [#929][])
 - **amaru-ledger**: introduce `StakePoolCostTooLowPOOL` coverage. ([#1037][], [#909][])
 - **amaru-consensus**: add events and metrics to track the performance of headers processing. ([#1005][])
+- **amaru-ledger**: benchmarks for key volatile db operations (roll forward, switch to fork and context preparation).
 
 ### Changed